+++ doc/includes/CHANGELOG.md
@@ -4,9 +4,9 @@
 
 ### What's new
 
-  * All queries are now started through `Model.query` [#346](https://github.com/Vincit/objection.js/issues/346)
-  * Objection is no longer transpiled using Babel. One of the implications is that you can use a objection github
-    link in package.json to test experimental version.
+  * All queries now call `Model.query` to create a `QueryBuilder` instance [#346](https://github.com/Vincit/objection.js/issues/346)
+  * Objection is no longer transpiled using Babel. One of the implications is that you can use a github
+    link in package.json to test experimental versions.
   * `count` can now be called without arguments [#364](https://github.com/Vincit/objection.js/issues/364)
 
 ### Breaking changes